moves utils for aixpax in routes/airpax
