+++ src/routes/airpax/TaskLists/TaskListSectionBuilder.jsx
@@ -4,7 +4,7 @@ import { Link } from 'react-router-dom';
 import { INDIVIDUAL_ICON } from '../../../constants';
 
 // Utils
-import { BaggageUtil, DateTimeUtil, IndicatorsUtil, BookingUtil, DocumentUtil, PersonUtil, MovementUtil } from './utils/index';
+import { BaggageUtil, DateTimeUtil, IndicatorsUtil, BookingUtil, DocumentUtil, PersonUtil, MovementUtil } from '../utils';
 import calculateTimeDifference from '../../../utils/calculateDatetimeDifference';
 
 const renderModeSection = (targetTask) => {
