+++ src/routes/airpax/__tests__/TaskListCard.test.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
 
-import TaskListCard from '../TaskListCard';
-import targetTask from '../../__fixtures__/targetListData';
+import TaskListCard from '../TaskLists/TaskListCard';
+import targetTask from '../__fixtures__/targetListData';
 
 describe('TaskListCard', () => {
   const airlineCodesMin = [
