+++ src/routes/airpax/utils/indicatorsUtil.jsx
